pyproject.toml
[tool.poetry]
name = "language-graph"
version = "0.1.0"
description = ""
authors = ["Spencer Peterson <spencer9696@gmail.com>"]
readme = "README.md"
packages = [{include = "language_graph"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.23.3"
jupyter = "^1.0.0"
plotly = "^5.15.0"
beautifulsoup4 = "^4.12.2"
selenium = "^4.10.0"
pandas = "^2.0.3"
numpy = "^1.25.0"
requests = "^2.31.0"
html5lib = "^1.1"
spacy = "^3.5.4"
pytest = "^7.4.0"
webdriver-manager = "^4.0.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
